+++ CHANGELOG.md
@@ -6,6 +6,7 @@ __New Feature__:
 - Add Project diffs
 - Import existing BigQuery Datasets and Tables
 - Generate Dataset and Table statistics
+- upsert table description
 
 # 0.6.3
 __New Feature__: